backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
scipy==1.13.0
numpy==1.26.4